LayoutTests/fast/history/history-replace-illegal-expected.txt

CONSOLE MESSAGE: line 9: SECURITY_ERR: DOM Exception 18: An attempt was made to break through the security policy of the user agent.
This page should trigger an error on attempting to set the URL bar to https://www.test.com/
